backend/requirements/common.txt
django~=3.0.14
djangorestframework
djangorestframework-jwt
django-cors-headers
django-pydenticon